Retry one more time if ratslap.UnknownRatslapError is raised
Sometimes ratslap.UnknownRatslapError is raised if the user runs the
program without mouse connected and plugs in the mouse afterwards.
This is due to mouse_state_changed signal is emitted before udev
finishes its job on mouse. So waiting for some time (100 msecs) and
trying again is a good idea.

Also when a device is plugged in, several device actions can be read
from usb_detector.USBDetector's monitor such as bind, add etc. Bind
action always comes after add action so searching for a bind action
is also a good idea if we want to delay the mouse_state_changed signal
from emitting.